created UseWindowWidth component fixing bugs in Header and CardsList
