+++ src/components/UseWindowWidth.jsx
@@ -1,0 +1,21 @@
+"use client";
+
+import { useEffect, useState } from "react";
+
+export default function UseWindowWidth() {
+  const [windowWidth, setWindowWidth] = useState(0);
+
+  useEffect(() => {
+    const handleResize = () => {
+      setWindowWidth(window.innerWidth);
+    };
+
+    window.addEventListener("resize", handleResize);
+    handleResize();
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
+  }, []);
+
+  return windowWidth;
+}
